requirements.txt
chromadb
fastapi
python-dotenv
uvicorn
